# server/rest/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /records/{namespace}/{record.resource}/_bulk:
        put:
            tags:
                - RecordService
            operationId: RecordService_Update
            parameters:
                - name: namespace
                  in: path
                  required: true
                  schema:
                    type: string
                - name: record.resource
                  in: path
                  required: true
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
                - name: record.id
                  in: query
                  schema:
                    type: string
                - name: record.resource
                  in: query
                  schema:
                    type: string
                - name: record.dataType
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: record.auditData.createdOn.seconds
                  in: query
                  description: Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z inclusive.
                  schema:
                    type: integer
                    format: int64
                - name: record.auditData.createdOn.nanos
                  in: query
                  description: Non-negative fractions of a second at nanosecond resolution. Negative second values with fractions must still have non-negative nanos values that count forward in time. Must be from 0 to 999,999,999 inclusive.
                  schema:
                    type: integer
                    format: int32
                - name: record.auditData.updatedOn.seconds
                  in: query
                  description: Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z inclusive.
                  schema:
                    type: integer
                    format: int64
                - name: record.auditData.updatedOn.nanos
                  in: query
                  description: Non-negative fractions of a second at nanosecond resolution. Negative second values with fractions must still have non-negative nanos values that count forward in time. Must be from 0 to 999,999,999 inclusive.
                  schema:
                    type: integer
                    format: int32
                - name: record.auditData.createdBy
                  in: query
                  schema:
                    type: string
                - name: record.auditData.updatedBy
                  in: query
                  schema:
                    type: string
                - name: record.version
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: checkVersion
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateRecordResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - RecordService
            operationId: RecordService_Create
            parameters:
                - name: namespace
                  in: path
                  required: true
                  schema:
                    type: string
                - name: record.resource
                  in: path
                  required: true
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
                - name: record.id
                  in: query
                  schema:
                    type: string
                - name: record.resource
                  in: query
                  schema:
                    type: string
                - name: record.dataType
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: record.auditData.createdOn.seconds
                  in: query
                  description: Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z inclusive.
                  schema:
                    type: integer
                    format: int64
                - name: record.auditData.createdOn.nanos
                  in: query
                  description: Non-negative fractions of a second at nanosecond resolution. Negative second values with fractions must still have non-negative nanos values that count forward in time. Must be from 0 to 999,999,999 inclusive.
                  schema:
                    type: integer
                    format: int32
                - name: record.auditData.updatedOn.seconds
                  in: query
                  description: Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z inclusive.
                  schema:
                    type: integer
                    format: int64
                - name: record.auditData.updatedOn.nanos
                  in: query
                  description: Non-negative fractions of a second at nanosecond resolution. Negative second values with fractions must still have non-negative nanos values that count forward in time. Must be from 0 to 999,999,999 inclusive.
                  schema:
                    type: integer
                    format: int32
                - name: record.auditData.createdBy
                  in: query
                  schema:
                    type: string
                - name: record.auditData.updatedBy
                  in: query
                  schema:
                    type: string
                - name: record.version
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: ignoreIfExists
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateRecordResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /records/{namespace}/{resource}:
        get:
            tags:
                - RecordService
            operationId: RecordService_List
            parameters:
                - name: namespace
                  in: path
                  required: true
                  schema:
                    type: string
                - name: resource
                  in: path
                  required: true
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  schema:
                    type: integer
                    format: uint64
                - name: useHistory
                  in: query
                  schema:
                    type: boolean
                - name: resolveReferences
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListRecordResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /records/{namespace}/{resource}/_bulk:
        put:
            tags:
                - RecordService
            operationId: RecordService_Delete
            parameters:
                - name: namespace
                  in: path
                  required: true
                  schema:
                    type: string
                - name: resource
                  in: path
                  required: true
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
                - name: id
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
                - name: ids
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteRecordResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /records/{namespace}/{resource}/_search:
        post:
            tags:
                - RecordService
            operationId: RecordService_Search
            parameters:
                - name: namespace
                  in: path
                  required: true
                  schema:
                    type: string
                - name: resource
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SearchRecordRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SearchRecordResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /records/{namespace}/{resource}/{id}:
        get:
            tags:
                - RecordService
            operationId: RecordService_Get
            parameters:
                - name: namespace
                  in: path
                  required: true
                  schema:
                    type: string
                - name: resource
                  in: path
                  required: true
                  schema:
                    type: string
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetRecordResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
        AuditData:
            type: object
            properties:
                createdOn:
                    type: string
                    format: date-time
                updatedOn:
                    type: string
                    format: date-time
                createdBy:
                    type: string
                updatedBy:
                    type: string
...
        BooleanExpression:
            type: object
            properties:
                and:
                    $ref: '#/components/schemas/CompoundBooleanExpression'
                or:
                    $ref: '#/components/schemas/CompoundBooleanExpression'
                not:
                    $ref: '#/components/schemas/BooleanExpression'
                equal:
                    $ref: '#/components/schemas/PairExpression'
                lessThan:
                    $ref: '#/components/schemas/PairExpression'
                greaterThan:
                    $ref: '#/components/schemas/PairExpression'
                lessThanOrEqual:
                    $ref: '#/components/schemas/PairExpression'
                greaterThanOrEqual:
                    $ref: '#/components/schemas/PairExpression'
                isNull:
                    $ref: '#/components/schemas/Expression'
                regexMatch:
                    $ref: '#/components/schemas/RegexMatchExpression'
        CompoundBooleanExpression:
            type: object
            properties:
                expressions:
                    type: array
                    items:
                        $ref: '#/components/schemas/BooleanExpression'
        CreateRecordResponse:
            type: object
            properties:
                record:
                    $ref: '#/components/schemas/Record'
                records:
                    type: array
                    items:
                        $ref: '#/components/schemas/Record'
                error:
                    $ref: '#/components/schemas/Error'
                inserted:
                    type: array
                    items:
                        type: boolean
...
        DeleteRecordResponse:
            type: object
            properties:
                error:
                    $ref: '#/components/schemas/Error'
...
        Error:
            type: object
            properties:
                code:
                    type: integer
                    format: enum
                message:
                    type: string
                fields:
                    type: array
                    items:
                        $ref: '#/components/schemas/ErrorField'
        ErrorField:
            type: object
            properties:
                recordId:
                    type: string
                property:
                    type: string
                message:
                    type: string
        Expression:
            type: object
            properties:
                property:
                    type: string
                functionCall:
                    $ref: '#/components/schemas/FunctionCall'
                value:
                    $ref: '#/components/schemas/GoogleProtobufValue'
        FunctionCall:
            type: object
            properties:
                package:
                    type: string
                name:
                    type: string
                arguments:
                    type: array
                    items:
                        $ref: '#/components/schemas/Expression'
        GetRecordResponse:
            type: object
            properties:
                record:
                    $ref: '#/components/schemas/Record'
                error:
                    $ref: '#/components/schemas/Error'
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        GoogleProtobufValue:
            description: Represents a dynamically typed value which can be either null, a number, a string, a boolean, a recursive struct value, or a list of values.
        ListRecordResponse:
            type: object
            properties:
                total:
                    type: integer
                    format: uint32
                content:
                    type: array
                    items:
                        $ref: '#/components/schemas/Record'
                error:
                    $ref: '#/components/schemas/Error'
...
        PairExpression:
            type: object
            properties:
                left:
                    $ref: '#/components/schemas/Expression'
                right:
                    $ref: '#/components/schemas/Expression'
        Record:
            type: object
            properties:
                id:
                    type: string
                resource:
                    type: string
                dataType:
                    type: integer
                    format: enum
                properties:
                    type: object
                references:
                    type: object
                auditData:
                    $ref: '#/components/schemas/AuditData'
                version:
                    type: integer
                    format: uint32
        RegexMatchExpression:
            type: object
            properties:
                pattern:
                    type: string
                expression:
                    $ref: '#/components/schemas/Expression'
...
        SearchRecordRequest:
            type: object
            properties:
                token:
                    type: string
                namespace:
                    type: string
                resource:
                    type: string
                query:
                    $ref: '#/components/schemas/BooleanExpression'
                limit:
                    type: integer
                    format: uint32
                offset:
                    type: integer
                    format: uint64
                useHistory:
                    type: boolean
                resolveReferences:
                    type: boolean
        SearchRecordResponse:
            type: object
            properties:
                total:
                    type: integer
                    format: uint32
                content:
                    type: array
                    items:
                        $ref: '#/components/schemas/Record'
                error:
                    $ref: '#/components/schemas/Error'
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        UpdateRecordResponse:
            type: object
            properties:
                record:
                    $ref: '#/components/schemas/Record'
                records:
                    type: array
                    items:
                        $ref: '#/components/schemas/Record'
                error:
                    $ref: '#/components/schemas/Error'
...